# internal/api/spec/knot-openapi.yaml
openapi: 3.0.3
servers:
  - url: https://127.0.0.1
    description: Local server.
info:
  version: 1.0.0
  title: Knot API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  x-logo:
    url: ../images/logo-light.svg
  description: |
    # Introduction
    The REST API is used for the CLI, Web interface and 3rd party integrations.
...
paths:
...
  /api/tunnels/{tunnel_name}:
    delete:
      tags:
        - Tunnels
      summary: Delete Tunnel
      description: Delete a specific tunnel for the authenticated user.
      operationId: deleteTunnel
      parameters:
        - name: tunnel_name
          in: path
          required: true
          description: The name of the tunnel to delete.
          schema:
            type: string
      responses:
        '200':
          description: Tunnel deleted successfully.
        '401':
          $ref: '#/components/responses/unauthorized'
        '404':
          $ref: '#/components/responses/not-found'
        '500':
          $ref: '#/components/responses/internal-server-error'
      security: [BearerAuth: []]
  /api/tunnels:
    get:
      tags:
        - Tunnels
      summary: Get Tunnels
      description: Retrieve all tunnels for the authenticated user.
      operationId: getTunnels
      responses:
        '200':
          description: A list of tunnels.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TunnelInfo'
        '401':
          $ref: '#/components/responses/unauthorized'
        '500':
          $ref: '#/components/responses/internal-server-error'
      security: [BearerAuth: []]
...
  /api/auth/web:
    post:
      tags:
        - Authorization
      summary: Login Web User
      description: |
        Authenticate a user with the server and create a web access token.

        A session is also opened and session cookie set within the response.
      operationId: WebLoginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: User authenticated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthLoginResponse'
        '400':
          $ref: '#/components/responses/bad-request'
        '401':
          $ref: '#/components/responses/unauthorized'
        '500':
          $ref: '#/components/responses/internal-server-error'
      security: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    AuthLoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: user@example.com
          description: The email address of the user
        password:
          type: string
          example: password
          description: The password of the user
        totp_code:
          type: string
          example: 123456
          description: The TOTP code for the user, ignored if TOTP is not enabled.
      required:
        - email
        - password
    AuthLoginResponse:
      type: object
      properties:
        status:
          type: boolean
          description: The status of the operation, true if successful login.
        token:
          type: string
          description: The token for the user session.
        totp_secret:
          type: string
          description: The TOTP secret for the user, if set then the code needs to be provided to the user as 1st login enables TOTP.
...
    TunnelInfo:
      type: object
      properties:
        name:
          type: string
          description: The name of the tunnel.
        address:
          type: string
          description: The address of the tunnel.
...
    error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: The error message.
...
  responses:
    bad-request:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
...
    not-found:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
    internal-server-error:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
...
    unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
...